+++ src/main/java/com/jovan/erp_v1/service/IInventoryItemsService.java
@@ -1,5 +1,6 @@
 package com.jovan.erp_v1.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.jovan.erp_v1.request.InventoryItemsRequest;
@@ -10,7 +11,7 @@ public interface IInventoryItemsService {
 	InventoryItemsResponse create(InventoryItemsRequest request);
 	InventoryItemsResponse update(Long id, InventoryItemsRequest request);
 	void delete(Long id);
-	List<InventoryItemsResponse> getByQuantity(Double quantity);
+	List<InventoryItemsResponse> getByQuantity(BigDecimal quantity);
 	List<InventoryItemsResponse> getByItemCondition(Integer itemCondition);
 	List<InventoryItemsResponse> getByInventoryId(Long inventoryId);
 	List<InventoryItemsResponse> getByProductId(Long productId);
@@ -18,5 +19,5 @@ public interface IInventoryItemsService {
 	InventoryItemsResponse findById(Long id);
 	List<InventoryItemsResponse> findAllInventories();
 
-    List<InventoryItemsResponse> findItemsWithDifference(Double threshold);
+    List<InventoryItemsResponse> findItemsWithDifference(BigDecimal threshold);
 }
